Add cookieParser middleware to the Yoga server
To enable authentication we need our Yoga server to be able to handle parsed cookie data since we will be storing the JWT in a cookie. Note that the decision to use cookie storage for the JWT rather than localstorage is because cookies are passed along in every request automatically whereas localstorage is not. Using cookie storage allows us to seamlessly perform server-side rendering of authenticated sections of the code.